fix: PR improvements - removed interfaces
