+++ extensions/jwks/src/main/java/de/sovity/edc/extension/jwks/JwksExtension.java
@@ -15,8 +15,8 @@
 package de.sovity.edc.extension.jwks;
 
 import de.sovity.edc.extension.jwks.controller.JwksController;
-import de.sovity.edc.extension.jwks.controller.JwksJsonTransformerImpl;
-import de.sovity.edc.extension.jwks.jwk.VaultJwkFactoryImpl;
+import de.sovity.edc.extension.jwks.controller.JwksJsonTransformer;
+import de.sovity.edc.extension.jwks.jwk.VaultJwkFactory;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.spi.security.Vault;
 import org.eclipse.edc.spi.system.ServiceExtension;
@@ -47,8 +47,8 @@ public class JwksExtension implements ServiceExtension {
             monitor.warning(() -> "No vault alias provided for JWKS-Extension");
         }
         var controller = new JwksController(
-                new VaultJwkFactoryImpl(vault),
-                new JwksJsonTransformerImpl(),
+                new VaultJwkFactory(vault),
+                new JwksJsonTransformer(),
                 pemSecretAlias,
                 monitor);
         webService.registerResource(controller);
